# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nouh LifeOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /finance/transactions:
    get:
      operationId: listTransactions
      tags: [finance]
      summary: List financial transactions
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [income, expense]
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: month
          schema:
            type: string
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [finance]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /finance/transactions/{id}:
    put:
      operationId: updateTransaction
      tags: [finance]
      summary: Update a transaction
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionBody"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [finance]
      summary: Delete a transaction
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /finance/budgets:
    get:
      operationId: listBudgets
      tags: [finance]
      summary: List budget categories
      responses:
        "200":
          description: Budgets list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: createBudget
      tags: [finance]
      summary: Create a budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetBody"
      responses:
        "201":
          description: Created budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"

  # ── CONTENT ──────────────────────────────────────────────────────
  /content:
    get:
      operationId: listContent
      tags: [content]
      summary: List content items
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [raw_idea, researching, scripting, filming, editing, ready, published, repurpose, archive]
        - in: query
          name: platform
          schema:
            type: string
      responses:
        "200":
          description: Content list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentItem"
    post:
      operationId: createContent
      tags: [content]
      summary: Create a content item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContentBody"
      responses:
        "201":
          description: Created content item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
  /content/{id}:
    get:
      operationId: getContent
      tags: [content]
      summary: Get a content item by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Content item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
    put:
      operationId: updateContent
      tags: [content]
      summary: Update a content item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContentBody"
      responses:
        "200":
          description: Updated content item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
    delete:
      operationId: deleteContent
      tags: [content]
      summary: Delete a content item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /daily-plans:
    get:
      operationId: listDailyPlans
      tags: [daily-plans]
      summary: List daily plans
      parameters:
        - in: query
          name: date
          schema:
            type: string
      responses:
        "200":
          description: Daily plans list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyPlan"
    post:
      operationId: createDailyPlan
      tags: [daily-plans]
      summary: Create or update today's daily plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDailyPlanBody"
      responses:
        "201":
          description: Created daily plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPlan"
  /daily-plans/{date}:
    get:
      operationId: getDailyPlan
      tags: [daily-plans]
      summary: Get daily plan by date
      parameters:
        - in: path
          name: date
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Daily plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPlan"
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        date:
          type: string
        note:
          type: string
        createdAt:
          type: string
      required: [id, title, amount, type, category, date, createdAt]
    CreateTransactionBody:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        date:
          type: string
        note:
          type: string
      required: [title, amount, type, category, date]
    UpdateTransactionBody:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        date:
          type: string
        note:
          type: string
...
    Budget:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        limit:
          type: number
        spent:
          type: number
        remaining:
          type: number
        period:
          type: string
          enum: [monthly, yearly]
      required: [id, category, limit, spent, remaining, period]
    CreateBudgetBody:
      type: object
      properties:
        category:
          type: string
        limit:
          type: number
        period:
          type: string
          enum: [monthly, yearly]
      required: [category, limit, period]

    # ── CONTENT ──────────────────────────────────────────────────────
    ContentItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        hook:
          type: string
        script:
          type: string
        status:
          type: string
          enum: [raw_idea, researching, scripting, filming, editing, ready, published, repurpose, archive]
        platform:
          type: string
          enum: [youtube, instagram, tiktok, twitter, linkedin, blog]
        targetAudience:
          type: string
        painPoint:
          type: string
        cta:
          type: string
        dueDate:
          type: string
        publishDate:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, status, platform, tags, createdAt, updatedAt]
    CreateContentBody:
      type: object
      properties:
        title:
          type: string
        hook:
          type: string
        script:
          type: string
        status:
          type: string
          enum: [raw_idea, researching, scripting, filming, editing, ready, published, repurpose, archive]
        platform:
          type: string
          enum: [youtube, instagram, tiktok, twitter, linkedin, blog]
        targetAudience:
          type: string
        painPoint:
          type: string
        cta:
          type: string
        dueDate:
          type: string
        publishDate:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [title, platform]
    UpdateContentBody:
      type: object
      properties:
        title:
          type: string
        hook:
          type: string
        script:
          type: string
        status:
          type: string
          enum: [raw_idea, researching, scripting, filming, editing, ready, published, repurpose, archive]
        platform:
          type: string
          enum: [youtube, instagram, tiktok, twitter, linkedin, blog]
        targetAudience:
          type: string
        painPoint:
          type: string
        cta:
          type: string
        dueDate:
          type: string
        publishDate:
          type: string
        tags:
          type: array
          items:
            type: string
...
    DailyPlan:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        mits:
          type: array
          items:
            type: string
        intention:
          type: string
        energyLevel:
          type: integer
        reflection:
          type: string
        wins:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, date, mits, wins, createdAt, updatedAt]
    CreateDailyPlanBody:
      type: object
      properties:
        date:
          type: string
        mits:
          type: array
          items:
            type: string
        intention:
          type: string
        energyLevel:
          type: integer
        reflection:
          type: string
        wins:
          type: array
          items:
            type: string
      required: [date]